restructure go code and added air to watch and recompiler code again
