+++ .air.toml
@@ -1,0 +1,53 @@
+root = "."
+testdata_dir = "testdata"
+tmp_dir = "tmp"
+include = ["src/**/*.go"]
+
+[build]
+  args_bin = []
+  bin = "./tmp/main"  # ✅ Correct binary path
+  cmd = "go build -o ./tmp/main main.go"  # ✅ Explicitly build only `main.go`
+  delay = 1000
+  exclude_dir = ["assets", "tmp", "vendor", "testdata"]
+  exclude_file = []
+  exclude_regex = ["_test.go"]
+  exclude_unchanged = true  # ✅ Prevents re-running if files haven't changed
+  follow_symlink = false
+  full_bin = ""
+  include_dir = []
+  include_ext = ["go", "tpl", "tmpl", "html"]
+  include_file = []
+  kill_delay = "0s"
+  log = "build-errors.log"
+  poll = false
+  poll_interval = 0
+  post_cmd = []
+  pre_cmd = []
+  rerun = false  # ✅ Prevents double execution
+  rerun_delay = 500
+  send_interrupt = false
+  stop_on_error = true  # ✅ Stop execution if the build fails
+
+[color]
+  app = ""
+  build = "yellow"
+  main = "magenta"
+  runner = "green"
+  watcher = "cyan"
+
+[log]
+  main_only = true  # ✅ Only log the main process
+  silent = false
+  time = false
+
+[misc]
+  clean_on_exit = true  # ✅ Clean tmp files on exit
+
+[proxy]
+  app_port = 0
+  enabled = false
+  proxy_port = 0
+
+[screen]
+  clear_on_rebuild = true  # ✅ Clear screen for better visibility
+  keep_scroll = true
