+++ main.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"golang-poc/src/hello"
+)
+
+func main() {
+	hello.Hello()
+	// bill.CreateBill()
+	// todo.StartTodo()
+	// scrapper.WebScrapper()
+	// basic.Basic()
+}
