+++ src/hello/index.go
@@ -1,0 +1,7 @@
+package hello
+
+import "fmt"
+
+func Hello() {
+	fmt.Println("Hello World!")
+}
